% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onePopPropCI.R
\name{onePopPropCI}
\alias{onePopPropCI}
\title{Confidence Interval for One Population Proportion}
\usage{
onePopPropCI(x, n, confidence = 0.95)
}
\arguments{
\item{x}{Number of successes}

\item{n}{Sample size}

\item{confidence}{Desired confidence level (default is 0.95)}
}
\value{
The confidence interval for a population proportion
}
\description{
Creates a confidence interval for a population proportion
}
\examples{
onePopPropCI(x = 60, n = 100)
onePopPropCI(x = 47, n = 112, confidence = 0.90)
}
