% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomSample.R
\name{randomSample}
\alias{randomSample}
\title{Random Sample Mean Function}
\usage{
randomSample(data, column = 1, n)
}
\arguments{
\item{data}{Data set of interest (a data frame)}

\item{column}{Column number corresponding to the variable of interest from the data frame (default = 1)}

\item{n}{Sample size}
}
\value{
The random sample of size n and the corresponding sample mean
}
\description{
This function calculates a sample mean from a data set.
}
