% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomSampleProportion.R
\name{randomSampleProportion}
\alias{randomSampleProportion}
\title{Random Sample Proportion Function}
\usage{
randomSampleProportion(data, column = 1, n)
}
\arguments{
\item{data}{Data set of interest (a data frame)}

\item{column}{Column number corresponding to the variable of interest from the data frame (default = 1)}

\item{n}{Sample size}
}
\value{
The random sample of size n and the corresponding sample proportion
}
\description{
This function calculates a sample proportion from a data set.
}
