% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMean.R
\name{sampleMean}
\alias{sampleMean}
\title{Sample Mean}
\usage{
sampleMean(
  data,
  n,
  plot = FALSE,
  title = "Histogram of a Random Sample",
  xlab = NULL,
  bins = NA
)
}
\arguments{
\item{data}{Variable of interest (dataset$variable)}

\item{n}{Sample size}

\item{plot}{Plots a histogram of the data when set to TRUE}

\item{title}{Supply a string for the title of the histogram}

\item{xlab}{Supply a string for the x-axis label}

\item{bins}{Number of bins used in the histogram (default is 30)}
}
\value{
A random sample of size n, the corresponding sample mean, and an optional histogram of the data
}
\description{
Collects a random sample of n observations and computes the sample mean
}
\examples{
sampleMean(trees$Height, 10)
sampleMean(iris$Sepal.Length, n = 80, plot = TRUE, xlab = "Sepal Length (in cm)", bins = 15)
}
