% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingDist.R
\name{samplingDist}
\alias{samplingDist}
\title{Sampling Distribution for a Sample Mean Function}
\usage{
samplingDist(data, column = 1, n)
}
\arguments{
\item{data}{Data set of interest (a data frame)}

\item{column}{Column number corresponding to the variable of interest from the data frame (default = 1)}

\item{n}{Sample size}
}
\value{
A histogram of 10,000 sample means (representing the sampling distribution of the sample mean) generated from 10,000 random samples of size n
}
\description{
This function creates a sampling distribution for the sample mean. This sampling distribution is visualized using 10,000 sample means taken from random samples of size n.
}
