% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingDistProportion.R
\name{samplingDistProportion}
\alias{samplingDistProportion}
\title{Sampling Distribution for a Sample Proportion Function}
\usage{
samplingDistProportion(data, column = 1, n)
}
\arguments{
\item{data}{Data set of interest (a data frame)}

\item{column}{Column number corresponding to the variable of interest from the data frame (default = 1)}

\item{n}{Sample size}
}
\value{
A strip chart of 100 sample proportions (representing the sampling distribution of the sample proportion) generated from 100 random samples of size n
}
\description{
This function creates a sampling distribution for the sample proportion. This sampling distribution is visualized using 100 sample proportions taken from random samples of size n.
}
