% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMeans.R
\name{simMeans}
\alias{simMeans}
\title{Simulation of the Sampling Distribution of the Sample Mean}
\usage{
simMeans(data, mu = NA, sigma = NA, n, reps = 500, zoom = FALSE, xlim = NA)
}
\arguments{
\item{data}{Variable of interest (dataset$variable)}

\item{mu}{Instead of providing data, the population mean (mu) and population standard deviation (sigma) can be specified for a normal distribution}

\item{sigma}{Instead of providing data, the population mean (mu) and population standard deviation (sigma) can be specified for a normal distribution}

\item{n}{Sample size}

\item{reps}{Number of repetitions (default is 500)}

\item{zoom}{Logical argument (TRUE or FALSE) for determining the plot window}

\item{xlim}{Manually set the x-axis bounds using a vector (default is NA)}
}
\value{
A simulated sampling distribution of the sample mean
}
\description{
This function simulates the sampling distribution of the sample mean
}
\examples{
simMeans(data = trees$Height, n = 20)
simMeans(mu = 100, sigma = 10, n = 50, reps = 1000)
}
