% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simProportions.R
\name{simProportions}
\alias{simProportions}
\title{Simulation of the Sampling Distribution of the Sample Proportion}
\usage{
simProportions(
  p,
  n,
  reps = 100,
  threshold = NA,
  alternative = "two.sided",
  results = FALSE,
  zoom = TRUE,
  xlim = NA,
  plotresults = FALSE
)
}
\arguments{
\item{p}{True proportion}

\item{n}{Sample size}

\item{reps}{Number of repetitions (default is 100)}

\item{threshold}{Observed sample proportion used for calculating a simulated p-value}

\item{alternative}{A character string specifying the direction of the alternative hypothesis ("less", "greater", or "two.sided" (default))}

\item{results}{Prints out the simulated p-value results when set to TRUE}

\item{zoom}{Logical argument (TRUE or FALSE) for determining the plot window}

\item{xlim}{Manually set the x-axis bounds using a vector (default is NA)}

\item{plotresults}{Displays the simulated p-value results below the plot when set to TRUE}
}
\value{
A simulated sampling distribution of the sample proportion
}
\description{
This function simulates the sampling distribution of the sample proportion with the ability to calculated a simulated p-value.
}
\examples{
simProportions(p = 0.5, n = 20)
simProportions(p = 0.5, n = 50, threshold = 0.52, alternative = "greater", results = TRUE)
}
