% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tMultiplier.R
\name{tMultiplier}
\alias{tMultiplier}
\title{T Multiplier}
\usage{
tMultiplier(confidence, df)
}
\arguments{
\item{confidence}{Desired confidence level}

\item{df}{Degrees of freedom}
}
\value{
The t multiplier for the specified confidence level
}
\description{
Computes the t multiplier for the specified confidence level
}
\examples{
tMultiplier(confidence = 0.95, df = 35)
}
