% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoPopPropCI.R
\name{twoPopPropCI}
\alias{twoPopPropCI}
\title{Confidence Interval for a Difference in Two Population Proportions}
\usage{
twoPopPropCI(x1, n1, x2, n2, confidence = 0.95)
}
\arguments{
\item{x1}{Number of successes in sample 1}

\item{n1}{Sample size of sample 1}

\item{x2}{Number of successes in sample 2}

\item{n2}{Sample size of sample 2}

\item{confidence}{Desired confidence level (default is 0.95)}
}
\value{
The confidence interval for a difference in two population proportions
}
\description{
Creates a confidence interval for a difference in two population proportions
}
\examples{
twoPopPropCI(x1 = 55, n1 = 100, x2 = 62, n2 = 100)
twoPopPropCI(x1 = 47, n1 = 112, x2 = 39, n2 = 102, confidence = 0.90)
}
