% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoPopPropHT.R
\name{twoPopPropHT}
\alias{twoPopPropHT}
\title{Hypothesis Test for a Difference in Two Population Proportions}
\usage{
twoPopPropHT(x1, n1, x2, n2, alternative = "two.sided")
}
\arguments{
\item{x1}{Number of successes in sample 1}

\item{n1}{Sample size of sample 1}

\item{x2}{Number of successes in sample 2}

\item{n2}{Sample size of sample 2}

\item{alternative}{A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less"}
}
\value{
Hypothesis test information including the sample proportions, resulting test statistic, and corresponding p-value
}
\description{
Tests the hypotheses for a difference in two population proportions
}
\examples{
twoPopPropHT(x1 = 55, n1 = 100, x2 = 62, n2 = 100)
twoPopPropHT(x1 = 47, n1 = 112, x2 = 39, n2 = 102, alternative = "greater")
}
