% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zMultiplier.R
\name{zMultiplier}
\alias{zMultiplier}
\title{Z Multiplier}
\usage{
zMultiplier(confidence)
}
\arguments{
\item{confidence}{Desired confidence level}
}
\value{
The z multiplier for the specified confidence level
}
\description{
Computes the z multiplier for the specified confidence level
}
\examples{
zMultiplier(0.95)
zMultiplier(confidence = 0.99)
}
