% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onePopPropHT.R
\name{onePopPropHT}
\alias{onePopPropHT}
\title{Hypothesis Test for One Population Proportion}
\usage{
onePopPropHT(x, n, p0, alternative = "two.sided")
}
\arguments{
\item{x}{Number of successes}

\item{n}{Sample size}

\item{p0}{Null population proportion (from H0)}

\item{alternative}{A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less"}
}
\value{
Hypothesis test information including the sample proportion, resulting test statistic, and corresponding p-value
}
\description{
Test the hypotheses for a population proportion
}
\examples{
onePopPropHT(x = 55, n = 100, p0 = 0.50)
onePopPropHT(x = 47, n = 112, p0 = 0.4, alt = "greater")
}
