% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomSample.R
\name{randomSample}
\alias{randomSample}
\title{Random Sample}
\usage{
randomSample(data, n, bins, title, xlab)
}
\arguments{
\item{data}{Variable of interest (dataset$variable)}

\item{n}{Sample size}

\item{bins}{Bins used in histogram}

\item{title}{Title used in histogram}

\item{xlab}{Label used on x-axis of histogram}
}
\value{
A random sample of size n from the specified data
}
\description{
Collects a random sample from provided data
}
\examples{
randomSample(iris$Petal.Length, n = 100, bins = 30, title = "Histogram of Petal Length", xlab = "Petal Length (in cm)")
}
